cv/requirements.txt
mysql-connector-python
flask


requests==2.24.0
json2html==1.3.0
user-agents==2.2.0
pytz==2021.1
pandas==1.1.5
opencv-python==4.5.1.48
numpy==1.20.0


# Detection
Pillow==9.0.0
tensorflow==2.11.0
tensorflowjs==4.1.0
testresources ==2.0.1
future==0.18.2
onnx==1.12.0
cupy==11.4.0
seaborn==0.12.1
numba==0.48
tqdm==4.56.0
more-itertools==8.12.0
matplotlib==3.5.1


# crowd localization
yacs
easydict
scipy
# seaborn==0.12.2
# tqdm


# Pillow
# tensorflow
# tensorflowjs
# testresources 
# future
# onnx
# cupy
# seaborn
# numba
# tqdm
# more-itertools
# matplotlib

# torch
# torchvision